requirements.txt
Django==3.2.12
psycopg2-binary
djangorestframework
